Sacrifice some security on the RPi for easy of use
Ignore changed host keys, which would stop the entire job, and rather accept whatever is coming. We thus trust the hostname lookup, and assume that no bothers to try to MitM a camera dumping public images. We're thus vulnerable to ARP spoofing, DNS hijacking and someone physically in the route to bu.roms.no.